requirements.txt
langchain
langgraph
faiss-cpu
sentence-transformers
streamlit
pypdf
duckduckgo-search
tiktoken